Update to be usable in per-room logic calculation
Edited old logical info to match the current state of the randomizer logic. Calculating logic per-room should now match the default logic which assumed a set map. This can be used for the randomizer to edit the logic in ways previously impossible.

This involves adding IDs to item locations, fixing errors in several places, as well as some changes to object fields.

Also made several minor changes to enemy-randomizer-relevant tags.