Add enemy rando style, ensure divekickable enemies
Do Your Worst logic can require that an enemy is present to divekick to reach a progression location. When that preset is selected, enemy randomizer will now ensure that one of the enemy locations has a flying enemy, to help avoid potential problems.

Also added the option to pick the enemy randomizer logic separate from item logic, mostly as a request for Hard Mode randomizer runs, but it could also be used for other styles of enemy randomization in the future.