Add hard mode boss AI, Arthroverta cutscene skip
When Arthroverta is placed in a room with an upwards door, entering through the upwards door and triggering the cutscene will cause the player to fall back out of the door and leave the room.

Currently there is only one possible room where this can happen, and other items can be used to enter the room in any logical situation where this room can be accessed, but it is still better not to require this of the player, and will be necessary anyway in future rooms with upward doors and no convenient Magnes point nearby. So the cutscene will be disabled in only the rooms where it causes problems.